Simplify ESLint command
This simplified command not only uses fewer characters, but is also more
future-proof: it will run ESLint against any files that match patterns
specified in the "overrides" section of the ESLint config. For example,
if we add an "overrides" section for .ts files in the future, ESLint
will automatically lint them.